Export AI Scribe code from controller to a mixin
Avoid adding too much code in the controller
